+++ CHANGELOG.md
@@ -94,6 +94,8 @@
 
 - Added aesthetics for `Stairs` [#573](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/573).
 
+- Added `Volume`, `Stairs` aesthetic mapping and `stephistogram` analysis [#572](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/572).
+
 ## v0.8.12 - 2024-10-07
 
 - Added `legend` keyword in `visual` to allow overriding legend element attributes [#570](https://github.com/MakieOrg/AlgebraOfGraphics.jl/pull/570).